move registering of actions/kbs out of progress part
